discriminate between "cli" and "non-cli" (api) mode
When used via CLI, it is usually a reasonable behaviour to early-exit
upon errors with a "SystemExit" (terminating the entire process). This
is less so when used via API.

Represent the "mode" (called by CLI vs via API) in the ctx; Change
util.Failure's base class (raised by util.fail) to RuntimeError unless
called via CLI (in which case the old behaviour is retained).