new trait: component_descriptor
Add a new trait to pipeline definition contract. If declared, components
are required to expose an executable (named .ci/component_descriptor),
which is run in a step of the same name.

Said step has a predefined output (COMPONENT_DESCRIPTOR_DIR), into which
the step execution is required to place a valid component descriptor.
This is subsequently uploaded as a github release asset as part of
release job execution.