optional container_registry cfg attribute (img-prefixes)
In some cases, credentials are required for interacting with container
registries (e.g. for retrieving images from private container
registries).

As multiple container image configurations exist, a lookup is required
to determine an appropriate container cfg for a given image reference.
This is done by declaring an optional set of prefixes, indicating that
a container registry cfg is appropriate for any container image whose
reference has the same prefix.