refactor model: mv concourse to own package
As a first step of finally splitting the much-too-large collection of
all configuration model element types into separate modules, start with
concourse cfg and move it to an own module 'concourse'.

In order to not break deployed deserialisers, do not introduce
fully-qualified type names yet, but implement a lookup that simulates a
flat type hierarchy.