upload triages upon application "updates"
Protecode does not allow to "re-use" uploaded "applications", which we
intended to do in order to represent component versions. By doing so, we
intended to keep previous "triages".

As Protecode, however, will always create new applications if the
uploaded contents have changed (which, of course, is always the case for
new component versions), we need to "transport" existing triages to new
Protecode applications to achieve this semantics.