rm erroneous type check
Currently, the 'type' argument in ArgumentSpecs is used to derive a
human-readible specification of allowed values (by stating the expected
types). While it would be possible to check that specified values must
actually be "type declarations" of some sort (either by using an actual
type, or a generic type from the typing module), said checking would
require some more efforts in order to be done right. On the other hand,
said checking does not add a lot of relevant type checking to our code.
So remove it. This (as a side-effect) will also allow our code to run
with Python3.7.