steps/release: refactor release creation
change release creation so that instead of four steps we only have two:
(local) creation of commits in the repository and the creation of the
release on GitHub (also handles tagging).
Also contained is a change to how the state of the local repository is
managed: Initially, the working tree is set (effectively via reset --hard)
to the state of HEAD before release processing is begun. Afterwards, the
commits are created in sequence and HEAD is updated for each commit.
Finally, the latest commit is pushed.
All stashing logic that existed previously has been removed.