determine protocol used by git using new option
previously, the protocol was determined either by the caller or
hardcoded. We've added a parameter to GithubConfigs that
allows to specify the available protocols to use when interacting with
git with a preferred order. This allows us to handle special cases such
as a GitHub instance without SSH endpoint without additional
checks by the caller.