fix issue in release notes
with the latest change to the release trait release notes are not
correctly processed. This happens because the tagged commits are not
directly reachable from master anymore. Thus all release notes since the
most recent release commit on the release-branch are always fetched.

This commit aims to fix this by considering the parents of the release-
commits (who are still on the release branch).