fix purge_history
Purge-History determines the newest commit that is older than the
specified cut-off date, and creates a new cut-off commit as a
replacement. After the second time this is run, apparently, the
thus-newly created cut-off commit is found as new replace-commit, which
will effectively prevent the actual cut-off semantics from being
applied.