fix bug and add support for reloading the cfg_set in whd
When trying to render a pipeline for a new JobMapping that is not
yet part of the cfg_set the WHD is given on initialisation, the WHD
will now encounter a newly defined JobMappingNotFoundError. Use this
error as signal to attempt a single reload of the cfg_set.

Previously, no error was raised and a AttributeError was encountered
at a later time in processing