order by author-date
commits in git have two dates: commit and author date. The latter does
not change, e.g. upon history rewrites, as opposed to the first, which
does.

For the special-case of keeping a defined time-range of past commits, in
a linear commit history, we thus must order according to the
author-dates (which do not change), as opposed to commit-dates (which
will be updated w/ every history-rewrite, thus leading to a situation,
where the originally intended behaviour (purging old commits) is not
done (except for the first attempt, or when only run after longer than
the configured retention-range).