Add transport tool for component descriptor v2
This commit adds the version 2 of the tool originally
created by Christian Cwienk.

It adds the possibility to use the tool offline in airgap environment.

The tool only supports ociImage resource type for the moment.

I have also updated the unit tests and added new ones for the
downloaders and uploaders.

Signed-off-by: lcavajani <lcavajani@suse.com>
Co-authored-by: ccwienk <christian.cwienk@sap.com>