add first parts of oci-client
`containerregistry-ccwienk` is cumbersome to use, and - while it does a
good job w/ handling all different docker-style OCI-Artifacts, it is not
intended as a general-purpose, lowlevel-API-client for OCI Registries,
and requires quite some monkey-patching to use it for that.

Also, through its usage of `httplib2`, and its tendency to make it hard
to immediately work w/ the issued http-requests, it makes it very hard
to switch to a streaming-approach, or introduce custom caching.

Thus, it seems adequate to implement a custom oci-registry-api-client,
which leaves more flexibility, as imposed by special-case scenarios,
such as verbatim OCI-Artifact replication.