more kaniko-madness
Resort to caching of cfg-element-types, so a cfg-factory-lookup (which
uses inspection / dynamic type-lookup) will succeed, even after kaniko
decided to purge our poor build container.

Admittedly, this is a yet another very dirty hack, and the cleaner
approach might be to accept that the build container cannot be used for
anything after running a multistage-build in it, and simply pass the
created image-tar to a second step, which then handles the upload. OTOH,
I am feeling rather stubborn today..