make kaniko-blob-wrapper a filelike-obj
Since a) GCR does not support chunked uploads [0], and b) I am stubborn,
and refuse to extract files from kaniko's tarfiles, and c) GCR will also
refuse to accept chunked-http-requests (which would be the outcome of
simply using the chunk-generator from previous commit), do the obbvious
and - using the same abstraction - create a fake filelike-object which
will delegate to the underlying tarfile's fileobj.

This is subsequently (separate commit) passed as `data` for the
monolithing upload-rq.

[0]
https://cloud.google.com/container-registry/docs/pushing-and-pulling#push_the_tagged_image_to