resolve nonlocal cfg-elements using bootstrap-cfg-factory
Prepare for planned splitting of cfg-elements into multiple
repositories. Start using a "bootstrapping" cfg-factory (which ignores
non-local cfg-sources) and thus provides github-cfg-elements, which are
required for resolving other (non-local) cfg-elements.

In the ctx of gardener-cicd-clusters, this does not change a lot, as
within a given cicd-cluster, all cfg is passed as a single `dict` (from
secrets-server).

For local (dev) scenarios, this causes all non-local cfg-elements to be
resolved upon creation of cfg-factory (which will probably be unresired
in most cases, as it will tend to take rather long). Therefore, lazy
cfg-element-retrieval should be implemented.

For central secrets-server-replication, however, this should be
acceptable / resolution will be required anyway.