change malware-scan mail-sending behaviour
previously, the code would refuse sending emails if no malware was
found. Changed this, so that an email containing all scan-results
will _always_ be sent if the recipients are explicitly configured.
In the case of component-owner-notification, the previously stated
behaviour has been fixed, so that it will send emails _unless_ there
is no malware found (to reduce spam).
Also changed the subject of the email to something less alerting in
case no malware is found.