workaround (too?) strict validation in artifactory
Kaniko will in some cases create cfg-blobs where the amount of entries
in the `history` attr is smaller than the amount of fs-layer-blobs.

This leads to artifactory rejecting the manifest with an error message
like so:

```
{"errors":[{"code":"MANIFEST_INVALID","message":"manifest
invalid","detail":{"description":"Circuit Breaker Threshold Reached,
Breaking Operation. see log output for manifest details."}}]}
```

There are some bugs reports to be found on that behaviour, e.g.:

https://github.com/GoogleContainerTools/jib/issues/875

As a workaround, cp the last history-entry in case the amount of layers
exceeds the history-entries to prevent this behaviour from happening.