only notify pipeline-owners if there is a cfg-err
There are reasons for pipeline-renderings to fail that are not related
to erronrous pipeline-definitions. For example, there are sporadic
networking issues, and from time to time, there are also bugs in our
pipeline-template.

In thoses cases, it does not help pipeline-owners to be notified, so
suppress notifications in those cases.