skip release-notes-importing on absent sources
CNUDIE-Component-Descriptors do not necessarily require to contain a
source-definition (also, on a side-note, source(s) do not necessarily
need to be GitHub-hosted git-repositories).

In the ctx of automatic upgrade-PR-creation, a github repository is
merely required for release-notes-retrieval. So simply skip
release-notes-import in case there are no source(s) for a given
component.