only run PR jobs for PR-head-commits
old behaviour was triggering the PR for every version of a given
PR which resulted in a lot of overhead for PRs with many commits.
This changes the WHD to only trigger once - for the head-commit.

If the head-commit cannot be determined for any reason (e.g.: delivery
schema-change) a warning will be printed and the code will revert to the old
behaviour.