optionally publish CD during CD step
add new option to publish the component descriptor to a repository-
context after it has been created during the CD step. If the option
is not set, publication will only happen during release-step (for both
CTF and CDv2).
This was done to make our handling of CD and CTF more consistent
with each other.