run kaniko-build in chroot-env
Kaniko assumes to be executed with an (almost) empty filesystem. Also,
it assumes it may purge everything it sees. The way we run kaniko does
not fulfil those assumptions (which required us to implement several
workarounds, such as "rescueing" file we need after calling kaniko, or
"hiding" files that would otherwise influence some builds).

Run kaniko in a chroot-env instead to get rid of those issues (and the
related ugly workarounds/hacks).