Workaround ghcr registry and media type distribution (#633)
* Fixed: auth issue with ghcr and throw NotImplemented if media type is application/vnd.docker.distribution.manifest.list.v2+json

GHCR OAuth flow does not return an JWT Token but a base64(username:password) that should be used. From this token, the expiration time could not be calcuated and we therefore set it to a fixed value.
Additionally, we have not implemented the handling of application/vnd.docker.distribution.manifest.list.v2+json, so we throw an error.