force kaniko-build (required if running on containerd)
kaniko checks for hints on whether it is executed within a
virtualisation container. This check seems to erroneously lead kaniko
into believing it is not running in a virtualisation container when
using containerd as container runtime for concourse (probably because we
mv away some marker files it relies on as a fallback).

Since we always do run our build steps in containers, tell kaniko to do
as we tell it.