revert Grafeas-client version change
due to a recent change in google's discovery-client the definitions
for the clients will be looked up in a local discovery-cache that
seems to be shipped with the package. At the moment this cache is
missing the definitions for the 'v1' containeranalysis client.
Disable static discovery, forcing the client to look it up every-
time the client is instantiated (this was the previous behaviour)