optionally run update-pr-step's callback in container
In some cases, it is desirable to perform non-trivial operations as part
of an upgrade-PR-content (for example, running a go-revendoring). To do
this, technology-specific tooling is required, which cannot reasonably
be integrated into the default job image.

Therefore, offer the option to configure a custom container image. If
configured, the callback is run in a container rather than a subprocess
in the default job container.

Note that authentication for image-pull is not yet in place. Also note
that there is no caching in place (so each upgrade-run will cause a full
image-pull).