allow meta-step to output multiple lines
We recently increased the amount of data contained in `meta`-step's
output. This will sometimes lead Concourse's buildlog-retrieval API
endpoint to split the output into multiple chunks. Thus, it is no longer
adequate to look for one single line. Collect the full log and parse it,
instead. This has the downside that any additional output might cause
parsing to fail.