omit referenced components when ordering CDs
cnudie.util.iter_sorted is intended to return a sequence of
component-descriptors with potential references between each other to
thus find the right deployment order to always ensure there be no
dangling references in the tgt component repository.

Component references outside the given set of component-descriptors
should thus not be included in the ordered sequence.