do not try to replicate cfg-blob if converting v1->v2
If replicating an OCI Image where the source has a legacy (v1) Manifest,
it is mandatory to convert the manifest to v2, as contemporary OCI
Registry implementations will not accept a v1-manifest.

As a consequence, the v2-manifest-blob (which was part of the v1-style
manifest) needs to be synthesised (and uploaded to tgt).

While this worked also w/o this change, the previous implementation
always tried to retrieve the cfg-blob from source, even if doing a
v1->v2-conversion, emitting a confusing/misleading warning in the
process (however ultimately successfully replicating the artifact).

In this case, no longer try to retrieve the cfg-blob from source (which
will never exist) to both save a http-request, and silence the warning.