fix: write blob to temporary file during v1->v2-conversion
oci-lib tries to do as much streaming as possible, avoiding both reading
(potentially large) chunks of data into memory, and also avoiding to
"detour" data through local file system (the latter in order to slightly
reduce roundtrip-times).

There is, however, a corner-case which will not (easily) work well if
target OCI registry is a GCR (google-container-registry), if doing a
manifest-conversion from v1->v2, if source blobs are gzip-compressed.

In this particular special-case, layer-blobs need to be uncompressed, so
their uncompressed data can be hashed (required for cfg-blob-data).
To do this, the source stream is wrapped in a generator. However,
uploading from a thus-wrapped generator will not work well w/ requests.
Also, chunked uploads (which would work) are not implemented by GCR.

Considering that this is a rare special-case (v1-manifests will fade-out
over time anyhow), write blobs to temporary file and accept the involved
performance penalty.