implement multiarch build and publishing
If OciBuilder is set to the (new) `docker-buildx`, and one or more
target-platforms are defined, run each platform-specific build in a
separate job. Add platform-identifier as suffix to target-tag.

As last operation, collect all variant-images (or, more specifically,
their oci-manifests' digests), and publish an "manifest-list" the
configured push-target.

Configuring build to target multiple platforms should be
backwards-compatible for all OCI-Clients that can properly work w/
manifest-lists (this is known to be true for docker and containerd).