support case of pipeline without main-repo for notifications
We want to add support for notifications to our replication-
pipelines, but these are rendered without a main-repository.
Since it does not make sense to arbitrarily define one, this commit
changes it so that the notification-template can handle cases
where the actual notification-sending might succeed. In case the
sending cannot ever succeed, an exception will be raised at render-time.

All these cases should only ever happen for pipelines that are _not_
rendered using the central replication.