restructure failure and abort-handling
roll the failure-publishing in to 'on_failure'/'on_abort'
tasks, since the previous 'ensure' was basically just a
combination of the two.
In effect this means that publishing will now happen
whether the job was aborted or failed (guaranteed via 'ensure').

There is some duplication happening here, but that is due to Mako-
functions being painfully difficult to work with in respect
to function-indentation.