always comment if pipeline-validation fails for PRs
previously, we used a label to ensure that we comment only once if
we detect a broken pipeline-definition (and once more as soon as it
is fixed). With this commit, we will comment to the pull-request
every time a change is pushed for as long as the definition is broken.
As before, we use a label to track the broken state will also add
a comment indicating that the pipeline-def is fixed once it _is_
fixed.