don't consider finished jobs when triggering PRs
with this commit, we will no longer consider finished (i.e.
not running) jobs when determining whether to trigger a job
for a pull-request.
As a result, merely labeling a PR with the required label (usually
'reviewed/ok-to-test') should be enough to trigger another job
to run - no need to manually pin and/or refresh resources.