fix template-path-calc in cli
the previous calculation fails when called via 'gardener-ci' after
installing our cli-package, as the package structure differs in
that case.
Use the more resilient concourse.paths, as the cli pypi-package
depends on that anyways.
Also simplify the managed paths down to 'template_include_dir' as
the template retriever will search the directory anyways.