pass resolved issue-labels, rather than target-artefacts
Compliance-Tracking-Issues are related to a certain compliance issue
type, and are targeted for a certain artefact (or group of artefacts).
To be able to find existing Tracking-Issues based on type and artefacts,
labels are added.

Currently, related artefacts always were either OCM Sources, or OCM
Resources, thus functions creating or discovering thus-labelled
tracking-issues expected OCM Artefacts (and components) as arguments,
and calculated derived labels from those arguments.

Prepare for allowing objects other than OCM Artefacts to be tracked;
pull-up calculation of labels, and pass-in labels to the aforementioned
functions, instead.