workaround: do not remove manually assigned assignee (race-condition)
If an existing tracking-issue is updated manually (seen for assignee
assignment once), while automated update is running, this will lead to
"lost-update".

As manual assignments are quite likely for tickets where we cannot
automatically determine an assignee, as a workaround, never explicitly
set assignees to "no assignees", thus not overwriting manually assigned
assignees.

This will not generally resolve the aforementioned race-condition, but
avoid it in this special-case.