make branch part of the resource identifier
with the old code there is a corner case of two resources
having the same identifier even though they are different.
This happens if a pipeline def defines put-steps for multiple
branches of the same repository. For these resources, the
qualifier of the identifier is 'output' and since all other
attributes are identical (they differ only in the branch name)
a duplicate is detected and one is discarded. Later on we
handle these resources correctly though, resulting in an
error when setting the pipeline in Concourse - the resource
is present in the task, but not the resources-definitions.