include metadata about ClamAV in our malware-scan results
most importantly, include a timestamp indicating the age of
the virus-definitions used by the scan. This will be used in a
later commit to avoid re-scans if no new findings are possible
(similar to Protecode).

Also includes the signature-version (an increasing int) and the
unmodified version-str as reported by ClamAV.