allow image-specific multiplatform-builds
Image-builds configured via `publish` trait may declare some images to
be built only for a subset of platforms (for example just
`linux/x86_64`), in parallel to other images being built for multiple
platforms. This may be helpful, e.g. if multiplatform is not required
for all images, and thus porting efforts can be reduced.