add support to skip some malware scans
skipping behaviour is controlled the "age threshold" parameter.
The scan will be skipped unless the previous scans findings
were based on a virus definition database older than the
current by a given amount of days.

In addition, if there was _no_ change to the virus definition,
scans will always be skipped as no new findings are possible.