prefer "multiarch" images (avoid failures if registry is not grc-like)
Some OCI-Registries, such as GCR will always return OCI-Image-Manifests
(defaulting to linux/amd64) if `accept`-header is omitted, even if the
underlying manifest is actually of image-index type. Not all registries
do that, though, and will return http-404 instead if such a manifest
exists, but accept-header is not set.

Thus, always set accept-header to avoid erroneously assuming an
image/tag does not exist.