add additional validation to release-note generation
check that the relation between versions is as expected if
they are passed explicitly. This is due to a recent change
allowing for the creation of downgrade-prs for which the
current code would pass these versions with switched semantics.

In future we might consider creating actual "downgrade release-
notes", but that will require some more work to make the code
aware of it in other places, for example in the release-trait