fix: cli.py cd edit with editor=vi for darwin
When used on darwin, no changes where detected but only with editor=vi.
Instead of editing the file directly, vi deletes and (re)creates
files on `:write`.
Said behaviour can be confirmed with `inotifywait` (for linux) or
`fswatch` (for darwin).
A workaround is to configure write/backup settings, especially setting
`nowritebackup` flag.
However, as this is vi's default behaviour, rather re(open) filehandle
as, effectively, a new file is created.