filter some release notes
with this commit, blacklisted release-note-blocks will now be
explicitly filtered from the release-note-blocks to be included.
Previously, blacklisted blocks were calculated, but never
considered afterwards.
Also updates the path relevant for draft-release-creation to
supply the relevant commit-information for this behaviour.
Finally, restructure the flow a bit so that blocks are handled
consistently and so that fewer redundant args are passed around.