Do not undo changes with merging (back) release-commit
There is the chance of reverting changes with merging back
release-commit, if there was a head-update *after* release-commit was
created (to tag) already.
Prior to creating merge-commit, head was fetched to reduce risks of
conflicts. Rather fail merging release-commit than risking unwanted
changes on default branch.

Should be reintroduced with consideration of fetch + rebase.