use unprocessed version in draft-release (again)
currently, the greatest version with matching minor must
be greater than the version passed in. This is usually
not the case for the versions handled during draft-releases
as they are finalised and thus ahead of the released versions.

In my opinion, the greatest version with matching minor should
not be constrained in this way by the reference version,
but that is a change for another commit since the function
is used in several places.