fix: Do not revert submodule updates on release
Prior to merging-back the release-commit, we rebase to apply latest
upstream changes. If these changes include a submodule change (e.g.
version bump via upgrade-pull-request), they are ultimately reverted.
This is caused by rebase not updating the submodule, but only the
`pointer` (`.git/modules/<name>/HEAD`). When creating merge-commit
with `index-to-commit` semantics, the submodule version from the
actual (not updated) submodule is taken and overwrites the `pointer`.