Adjust compliance issue closing comment
The fact "scanned element is no longer in BoM" is not true anymore
for issues with a digest that was not found because the finding could
also have been assessed. In that case the `latest_processing_date`
would turn to `None` and thereby the digest would change too. Therefore,
include this possibility in issue closing comment and, further, write
"assessed" in case there are still other findings of the same ocm
resource (because it is obviously still in BoM then).