allow specification of "extra-push-targets" in publish-trait
Built OCI Images (using publish-trait) can be published to additional
"push-targets" if specified via the `extra_push_targets` attribute.
Those push-targets can optionally contain a tag (in which case they are a
slight generalisation of `additional_tags`). If the tag is omitted, the
calculated tags are identical to those used for "main push target" (as
defined by `image` attribute). Additional push-targets (either with or
without tag) will in addition be published using a platform-specific
suffix in case of multiplatform-builds.