Add option to delete old BDBA products
Using this option, an interval X can be set after which "old" products
(i.e. products which were not (re-)scanned for the last X period) are
going to be deleted. Hence, this interval must not be set smaller than
the specified rescan interval since then the products are going to be
deleted continuously. Also, products belonging to resources which are
not scanned anymore (in no version) are not going to be deleted.