Update compliance snapshot state model class
A compliance snapshot state should be specific for an extension
service like bdba or the issue replication, however not for a
datatype since for example vulnerabilities and licenses are detected
in the same scan. Also, use enum instead of dataclass for statuses.

Note: Keep this backwards compatible for now and remove `datatype`
property once its usage is removed in the delivery-gear-extensions.