retain attributes from unknown `access`-types
OCM is designed as a typed, extensible model. For operations not
requiring the processing of certain aspects of model elements (e.g.
opaque replication of OCM Component Descriptors), it is the desired
behaviour for tooling to accept unknonwn model elements and retain them
unmodified.

Therefore, add `dict` as fallback deserialisation type for `dacite`. For
the sake of keeping code api-compatible, use custom `AccessDict` to
still expose attribute `type`, which is expected to be accessible as
attribute from existing users.