Redesign compliance scan data models
This redesign is intended to be more flexible with regards to adding
new scanners to the existing model. Therefore, an exchangeable id
is introduced which can differ based on the scanner, whereas the
actual scan data payload (e.g. the cve and severity for a vulnerability)
is specified consistently. This redesign currently only affects
vulnerability and license findings as well as structure info (formerly
known as components and filesystem paths).