Add component and reftype filter to `cnudie.iter`
The existing `node_filter` does not allow to filter all transitive
references of the filtered out nodes as well. Also, adding a flag
which enables such behaviour is not sufficient as there is the need
to have both kinds of filter in-place, e.g. to filter for resource
nodes but exclude the component `github.com/abc/xyz` and all it's
transitive references.

That's why, the component and reftype filter are being added. The
former allows to exclude certain components and all their transitive
references and the latter allows to exclude components and all their
transitive references based on the type how this component was
referenced by its parent, e.g. to exclude all so-called
"extra-component-references" which are referenced by label.