Add optional `references` to `ComponentArtefactId`
In case of runtime artefacts, the existence of a OCM component
descriptor `<component_name>:<component_version>` where the defined
artefact is referenced in does not have to exist. However, it may be
related to such a component or artefact (or both). To allow expressing
such semantics, a new field `references` is added, which allows defining
such relations in a recursive manner.