support files from build-steps as inlined ocm-release-assets
Add files (including directory trees) created from build-steps as
additional release-asset-type. For now, hardcode wrapping files in
xz-zipped tape-archive.

Configuration example (.ci/pipeline_definitions):

```
steps:
  step-with-output:
    out_dir: output-dir
traits:
  release:
    assets:
      - type: build-step-file
        name: my-asset-name
        step_name: step-with-output
        step_output_dir: output-dir
        path: '*'
```

In this example, step `step-with-output` is expected to create an
arbitrary amount of files and/or directories to its output directory
`output-dir`. Any directory contents will be included in release-asset
archive (unless name starts with `.`, as is posix globbing convention).

The resulting archive will be included as an inline-blob in the released
OCM-Component-Descriptor, and be referenced as a resource named
`my-asset-name`.