fix: allow adding input multiple times
e.g. if multiple release-assets are added, the same input might be added
multiple times. As inputs share a global namespace, this should not
actually be a reason for an error. Keep check for consistency of
variable/name-pair, to avoid inconsistent input-setting (where the
result will likely be wrong for all but the "winning" setter).