Factor-out manifest signing from manifest/blob upload
The separation of the actual signing of a manifest from interacting with
an OCI registry (i.e. for checking if a manifest already exists or
uploading it), eases testing the actual signing functionality as it does
not require interacting with an OCI registry (or at least a mock of
such) anymore.