fix: do not add publish-repository for notification-steps
"publish-repositories" are flavours of git-repo-resources that allow for
write-access, which are only exposed (and used) by build-steps
explicitly requesting that (via pipeline-definition).

They are special-handled (via resource-put). Such special-handling is
neither in place, nor desired for notification-steps (which are run on
error to send notifications, such as error-emails). However, declaring
output-steps as input for those notification-steps will lead to "missing
inputs: <resource-name>" errors if steps not configuring
publish-repository are executed (which happens only upon errors, and
therefore likely has slipped us for the past years.