cleanup: replace "NONE"-validation-mode w/ just None
It seems like a very quirky design-choice to have a validation-method,
that will early-exit if it is told to not do any validation. Instead,
callers should just not call this method, if no validation is desired.

Use builtin `None` to express the desire to omit validation, and have
validation-method consider it an error, if called w/o valid
ValidationMode.