refactor: remove dependencies from GitHelper to cfg-factory
Define new type `GitCfg`, which specifies minimal cfg (username +
email address, remote-url, credentials) for retrieving and / or
interacting with a git-repository with a remote.

In pipeline-template / code intended to be run within Gardener-CICD,
derive such GitCfgs from GitHubConfig.

This change will ultimately allow creation of GitHelpers also in
environments outside of Gardener-CICD (for example GitHub-Actions).

As GitUtil is used by `release_notes`, which is planned to be re-used in
GitHub-Actions, this is a preliminary step towards that.