add local component-cli to path
`cli.py pipeline component_descriptor` command was originally used for
migration from (deprecated) "component-cli" (golang binary) to
drop-in-replacement in this repository (bin/component-cli). Back then,
it made sense for the command to not implicitly add aforementioned
replacement to PATH.

With migration being done over a year ago, it appears adequate to make
the command more convenient, and let it add component-cli to PATH. Add
it to end of path to still allow callers to use local component-cli.