re-usable workflow for generating base-component-descriptor
"Base-Component-Descriptor(s)" have been proven to be beneficial for
integrating OCM-Component-Descriptors into CICD-Pipelines. Implement a
reusable workflow that will generate a base-component-descriptor to
avoid boilerplate in downstream pipelines.

This is a preliminary for enabling both a stand-alone
`component-descriptor` workflow/action(s), as well as support for
draft-release-pipelines (which can operate on
base-component-descriptors).