lookup release-notes in "parent" for forks
This will likely mostly be helpful for testing release-notes-action, as
draft-releases will typically not be of interest for forks, and releases
are typically not created for forked repositories. However, considering
it is fairly easy to determine parent, and will likely not interfere w/
vanilla-case, add it for the sake of convenience.

Caveat: we have a forked repository within gardener-org; for this
specific case, this behaviour will need to be overwritten.