add oci-ocm-multiarch workflow
With the advent of arm64-runners, it makes sense to have a convenient
reusable workflow for building multiarch OCI-Images (grouped in single
OCI Indexes), delegating image-builds to native runners (to avoid
overhead from emulation).

As runners can only be specified on job-level, this cannot be achieved
using an action (as done w/ ocm-oci-build action), hence use a reusable
workflow. Unfortunately, workflows are less flexible w.r.t.
customations, which results in a wider set of input-parameters.