fix: adjust documented artefact-names to implementation
Uploaded artefacts are named:

[ctx-]<digest>.ocm-artefacts (last part is separated w/ a period
character)

There are two obvious options at hand: Either adjust code to meet
documentation, or vice-versa (the latter is still an option, as there
should not yet be usage of this action).

Decide for the latter option to meet conventions also common for
filesystem-names, and to keep naming closer to contained
directory-tree-layouts.