refactor: rename input to just `component-descriptor`
It is not actually help- or meaningful to call the only input through
which a component-descriptor is to be passed to
merge-ocm-fragments-action _base_-component-descriptor. Also, such a
name might be misleading, as `base-component-descriptor` is (in
cc-utils-repository) thus-far used to describe component-descriptors
that are derived from metadata available to the pipeline. Merge-action
will accept _any_ component-descriptor as input.