fix: Install dependencies for `async` variant during lint
`gardener-oci` can optionally be installed in the asynchronous (async) variant.
As the linting scopes the whole repository, thus the async modules as
well, async related dependencies are required during lint.

Previously, this was given due to peer dependencies from `slackclient`.
However, with migration to `slack-sdk`, see:
https://github.com/gardener/cc-utils/commit/3fb4b93907cfa748ce54cb1bda29c4bd888bf11f
this peer dependency was dropped.