update documentation on head-update
We used to publish documentation-updates (for
Concourse-Pipeline-Template) as part of release, such as to avoid
publishing documentation about unreleased features of pipeline-contract.
However, development for Concourse-Pipeline-Template is mostly limited
to maintenance, and migration to GitHub-Actions is ongoing. Hence,
always publish documentation, which will allow us to roll-out
documentation related to GitHub-Actions in a more timely manner.