workaround: improve handling of OCI-Manifests w/o mediaType
Some few (we currently know of _one_) OCI-Manifest do not feature a
`mediaType` toplevel-attribute (as recommended by OCI-Spec). This will
lead to deserialisation errors, if the underlying manifest is not a
single image-manifest, but a Manifest-List / OCI-Image-Index.

To improve / make oci package more robust, honour Content-Type conveyed
from OCI-Registries via the `Content-Type` header in those cases.