limit size of commit-objects output from capture-commit action
release-commit is captured as part of `prepare.yaml` action such as to
already be able to include reference to it in base-component-descriptor
w/o pushing it (and of course to use it for release-action).

The captured commit objects are already consumed via GitHub-Artefact,
hence outputting it as base64-encoded string is not strictly necessary;
keep it for now, but limit it to 64 KiB (otherwise emit a warning, but
do not output it) to avoid errors due to limitation of total
output-size.