adjust documentation to reality
While originally, it was intended for update-component-trait to also
cover referenced resources (dependencies, called back when the
documentation was written, as back then, in "v1" (predecessor of OCM and
"CNUDIE"), references to resources or components were called so).
However, this was never implemented. With Concourse-Pipelines being
migrated to GitHub-Actions, this will also likely never happen.

In addition, it was originally intended to use the names of
component-refs or resources to communicate to callbacks what to do.
However, from the beginning, the component-names were passed instead
(likely, because back then (v1..), component-references (as opposed to
resource-references) did not _have_ names). For backwards-compatibility,
we will have to stick to defined env-var-name. However, at least adjust
documentation accordingly to reflect reality.